requirements.txt
Flask==3.0.0
openai-whisper==20231117
yt-dlp==2023.10.13
tqdm==4.66.1
Werkzeug==3.0.1
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
